pyproject.toml
[project]
name = "prompt-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.34.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
]
